+++ server/zanata-war/src/main/java/org/zanata/webtrans/client/presenter/AppPresenter.java
@@ -60,11 +60,12 @@ import com.google.gwt.event.logical.shared.ValueChangeHandler;
 import com.google.gwt.user.client.History;
 import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.rpc.AsyncCallback;
+import com.google.gwt.user.client.ui.Anchor;
 import com.google.gwt.user.client.ui.CheckBox;
 import com.google.gwt.user.client.ui.Widget;
 import com.google.inject.Inject;
 
-public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implements ValueChangeHandler<String>
+public class AppPresenter extends WidgetPresenter<AppPresenter.Display>
 {
    // javac seems confused about which Display is which.
    // somehow, qualifying WidgetDisplay helps!
@@ -102,7 +103,7 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
 
       void setUserLabel(String userLabel);
 
-      void setWorkspaceNameLabel(String workspaceNameLabel);
+      void setWorkspaceNameLabel(String workspaceNameLabel, String workspaceTitle);
 
       void setSelectedDocument(DocumentInfo document);
 
@@ -146,6 +147,8 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
    private final TranslationStats selectedDocumentStats = new TranslationStats();
    private final TranslationStats projectStats = new TranslationStats();
 
+   private static final String WORKSPACE_TITLE_QUERY_PARAMETER_KEY = "title";
+   
    private final UserOptionsPanel userOptionsPanel = new UserOptionsPanel(true, eventBus);
 
    @Inject
@@ -159,7 +162,6 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
       this.translationPresenter = translationPresenter;
       this.workspaceContext = workspaceContext;
    }
-
 
    @Override
    protected void onBind()
@@ -256,25 +258,23 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
          }
       }));
 
-      // registerHandler(display.getEditorButtonsCheckbox().addClickHandler(new
-      // ClickHandler()
-      // {
-      // @Override
-      // public void onClick(ClickEvent event)
-      // {
-      // boolean showButtons = ((CheckBox)
-      // display.getEditorButtonsCheckbox()).getValue();
-      // eventBus.fireEvent(new ButtonDisplayChangeEvent(showButtons));
-      // }
-      // }));
-
       display.setUserLabel(identity.getPerson().getName());
 
-      display.setWorkspaceNameLabel(workspaceContext.getWorkspaceName());
+      String workspaceTitle = Window.Location.getParameter(WORKSPACE_TITLE_QUERY_PARAMETER_KEY);
+
+      display.setWorkspaceNameLabel(workspaceContext.getWorkspaceName(), workspaceTitle);
 
       Window.setTitle(messages.windowTitle(workspaceContext.getWorkspaceName(), workspaceContext.getLocaleName()));
 
-      History.addValueChangeHandler(this);
+      History.addValueChangeHandler(new ValueChangeHandler<String>()
+      {
+
+         @Override
+         public void onValueChange(ValueChangeEvent<String> event)
+         {
+            processHistoryEvent(event);
+         }
+      });
 
       History.fireCurrentHistoryState();
    }
@@ -322,19 +322,16 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
       wordCount.decrement(Updateevent.getPreviousStatus(), Updateevent.getWordCount());
    }
 
-   @Override
-   public void onValueChange(ValueChangeEvent<String> event)
+   private void processHistoryEvent(ValueChangeEvent<String> event)
    {
 
       Log.info("Responding to history token: " + event.getValue());
 
       HistoryToken token = HistoryToken.fromTokenString(event.getValue());
 
-      DocumentId docId = token.getDocumentId();
+      DocumentId docId = documentListPresenter.getDocumentId(token.getDocumentPath());
 
-      // comparing longs here as DocumentId objects don't seem to compare
-      // properly
-      if (token.hasDocumentId() && (selectedDocument == null || selectedDocument.getId().getId() != docId.getId()))
+      if (docId != null && (selectedDocument == null || !selectedDocument.getId().equals(docId)))
       {
          Log.info("Firing document selection event");
          try
@@ -348,7 +345,13 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
          Log.info("Fired document selection event for " + docId.getId());
       }
 
-      if (token.hasView() && token.getView() != display.getCurrentView())
+      // if there is no valid document, don't show the editor
+      if (docId == null)
+      {
+         token.setView(MainView.Documents);
+      }
+
+      if (token.getView() != display.getCurrentView())
       {
          if (display.getCurrentView().equals(MainView.Editor))
          {
@@ -364,6 +367,13 @@ public class AppPresenter extends WidgetPresenter<AppPresenter.Display> implemen
          display.showInMainView(token.getView());
       }
 
+      // update toggle link with alternate view, or doc list if no doc is
+      // loaded
+      if (docId == null || token.getView().equals(MainView.Editor))
+         token.setView(MainView.Documents);
+      else
+         token.setView(MainView.Editor);
+      ((Anchor) display.getDocumentsLink()).setHref("#" + token.toTokenString());
    }
 
 }