+++ server/zanata-war/src/test/java/org/zanata/rest/service/ResourceTestUtil.java
@@ -3,6 +3,8 @@ package org.zanata.rest.service;
 import java.util.Iterator;
 import java.util.List;
 
+import org.zanata.rest.dto.extensions.gettext.HeaderEntry;
+import org.zanata.rest.dto.extensions.gettext.PoHeader;
 import org.zanata.rest.dto.extensions.gettext.PoTargetHeader;
 import org.zanata.rest.dto.extensions.gettext.TranslationsResourceExtension;
 import org.zanata.rest.dto.resource.AbstractResourceMeta;
@@ -55,6 +57,19 @@ class ResourceTestUtil
             }
          }
       }
+   }
+   
+   static void addPoHeader( Resource res, String key, String value )
+   {
+      PoHeader poHeader = res.getExtensions(true).findByType(PoHeader.class);
+      
+      if( poHeader == null )
+      {
+         poHeader = new PoHeader();
+         res.getExtensions().add(poHeader);
+      }
+      
+      poHeader.getEntries().add( new HeaderEntry(key, value) );
    }
 
 }